+++ glide2gl/src/Glide64/glide64_rdp.c
@@ -847,14 +847,14 @@ static void rdp_texrect(uint32_t w0, uint32_t w1)
       lr_x += 1.0f;
       lr_y += 1.0f;
    } else if (lr_y - ul_y < 1.0f)
-      lr_y = ceil(lr_y);
+      lr_y = ceilf(lr_y);
 
    if (settings.increase_texrect_edge)
    {
       if (floor(lr_x) != lr_x)
-         lr_x = ceil(lr_x);
+         lr_x = ceilf(lr_x);
       if (floor(lr_y) != lr_y)
-         lr_y = ceil(lr_y);
+         lr_y = ceilf(lr_y);
    }
 
    //*/
